03_machine_learning_models/02_plant_disease_recognition/requirements.txt
Pillow==11.3.0
timm==1.0.19
torch==2.5.1
torchvision==0.20.1
